backend/pyproject.toml
[project]
name = "dataharvest"
version = "0.1.0"
description = "AI-powered structured data extraction tool"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "openai>=1.50.0",
    "anthropic>=0.39.0",
    "httpx>=0.27.0",
    "playwright>=1.48.0",
    "openpyxl>=3.1.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
pdf-export = ["weasyprint>=62.0"]

[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"
